fixed json not parsing when used as jar
in jar vs ide the files work differently, so I needed a new way of
finding the recipe files in the jar. The new method is a lot more
complicated and a bit slower, but it must be this way. I will
work on making this more efficient, but it only runs on startup.