seats logic rework
seats now implement a 'control seat' system, where there are seats that can control the entity, and ones that can't.
This new system adhere's to 1.9+ compatibility for entity seat management.
This should also fix some rare bugs where seats end up in the void.
This also fixed a bug where seats don't respawn when the player logs out in a seat.

cleared a redundancy in the fear horn AI.