you're gonna want to read this one.
The prank is things actually got fixed.

- Bogie logic vastly improved, there's a lot less problems around sharp corners, a little less data to manage, and it's a little more performance friendly.
NOTE: Interaction with railcraft may be slightly delayed by a tick or two at speeds over 70km/h, but it shouldn't be game breaking.

- Acceleration rework. the new system runs off all the data that the old system did, but takes into consideration type efficiency, and has a semi-linear scale from stats, it's not overly realistic, but it feels a LOT more like TC4/TCCE, while still adding in all the new features like multi-heading, and making the handling of each train feel a bit more unique. This new system is also reasonably more performance friendly, not to mention simpler.

- Drag and slope logic have been combined into the same thing and are not fully handled by the rollingstock class (trains will inherit it).
The new calculations are again, more like TC4/TCCE but are also add in the new features like independent stock brakes and weights, so each stock/train/consist will move more uniquely.

- API CHANGE: `getEfficiency()` is now `getFuelEfficiency()`
this was only a name change for legibility sake. In the future we will be adding a running efficiency method which will be taken into account for trains, and I don't want to get these two confused.

- fixed an issue with trains and stock potentially colliding with themselves.
- improved roadkill logic (more efficient/reliable but probably needs a buff).

linking is still kinda busted. also sharp 90 degree turns yeet things a little, but it's not totally gameplay breaking, and it's still better than it was, so, well enough for now.