Beginning of 2D items
- Added support for rendering 2D blocks.
The function `getIconResource` returns the Resource Location for the texture, null if no 2D texture.
The method `force2dItem` can circumstantially force the item to render 2D even when 3d items are enabled.

- Train/stock items now 'support' 2D TCCE/TC4 style icons, this only works for the shay currently (picked by RNG), more will come next update.
Train/stock items will automatically fallback to the 3D render if no texture is found for the item.

- Fixed a large number of position bugs with items in world/hand/item frame/etc.
- Fixed several bugs with signal logic.