Seat Manager Stage 2
improvements to GUI centering.
added player head indicator for a visual of who is in the seat in addition to the player name as the button.

only known issue with it right now is *sometimes* the packets randomly get a nullpointer and I haven't found out where/how yet. Soon:tm:.