More refinement to seat GUI
two crash fixes; one for opening any inventory that doesn't have seats and one for spam swapping seats on server.

There is one more issue that needs fixed and will be the last:tm: push related to the seat GUI: when in a loco with more than one seat, after changing out of the main seat you can no longer open the GUI to change seats.